lowered numPoints in Kmean, set epsilon to 2
no memory problems with 10 points, try larger number of points. Tested
this on Arduino and the threshold appears to be working.

+++ Kmeans.h
@@ -23,7 +23,7 @@ private:
     byte lastCentroids[2];
     
     //algo params
-    short epsilon;
+    byte epsilon;
     int numPoints;
     
     //algo output
@@ -53,7 +53,7 @@ public:
     void changeNumPoints(int numPoints);
     
     //get threshold
-    short getThreshold();
+    byte getThreshold();
 };
 
 #endif /* defined(__LIT_Headband_Simulator__Kmeans__) */
